Introduced an encapsulating Module 'WikihouseExtension' to allow for namespace recalling between different files of ruby code.
Beginning to split up wikihouse.rb into smaller files of related code. 

Run flags and config constants now in wikihouse_extension_loader before
extension is loaded.

Extra constants now in constants.rb
Utility functions and classes now in utils.rb
Output writters now in writers.rb

SVG output currently has an error. Investigating.
